+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - Titles are now parsed only if they occur after the series token AND after
   either issue, year or volume.
+- A more sophisticated date parser.
 - Issue numbers that lead with a '#' character may start with alphabetical
   characters.
 - If volume is parsed, but issue number is not, the issue number is copied from